category rewrited [Backend]

+++ HeshmatNews/Services/CategoryService.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using AutoMapper;
 using HeshmastNews.Data;
 using HeshmastNews.DTOs;
 using HeshmastNews.Entities;
@@ -10,55 +11,78 @@ namespace HeshmastNews.Services
     public class CategoryService : ICategoryService
     {
         private ApplicationDbContext _context;
+        private IMapper _mapper;
 
-        public CategoryService(ApplicationDbContext context)
+        public CategoryService(ApplicationDbContext context, IMapper mapper)
         {
             _context = context;
+            _mapper = mapper;
         }
 
-        public List<Category> GetAllCategories()
+        public List<CategoryVewModelDTO> GetAllCategories()
         {
-            return _context.Categories.ToList();
-        }
-
-        public List<Category> GetAllSubCategory()
-        {
-          return  _context.Categories.Where(c=>c.ParentId==null)
-              .Include(c=> c.Categories)
-              .ToList();
-        }
-
-        public Category GetCategoryById(int categoryId)
-        {
-            return _context.Categories.Find(categoryId);
-        }
-
-
-        public Category AddCategory(CategoryCreationDTO category)
-        {
-            var CategoryCreationDTO = new Category()
+            var category = new List<CategoryVewModelDTO>();
+            var categoryDb = _context.Categories.ToList();
+            foreach (var item in categoryDb)
             {
-                CateGoryName = category.CateGoryName,
-                ParentId = category.ParentId
-            };
-            _context.Categories.Add(CategoryCreationDTO);
-            _context.SaveChanges();
-            return CategoryCreationDTO;
+                category.Add(_mapper.Map<CategoryVewModelDTO>(item));
+            }
+
+            return category;
         }
 
-        public Category UpdateCategory(int categoryId, CategoryCreationDTO category)
+        public List<CategoryVewModelDTO> GetAllParrentCategory()
         {
-            var categoryDb = GetCategoryById(categoryId);
-            categoryDb.CateGoryName = category.CateGoryName;
-            category.ParentId = category.ParentId;
-            _context.Categories.Update(categoryDb);
+            var category = new List<CategoryVewModelDTO>();
+            var categoryDb = _context.Categories
+                .Where(x => x.ParentId == null).ToList();
+            foreach (var item in categoryDb)
+            {
+                category.Add(_mapper.Map<CategoryVewModelDTO>(item));
+            }
+
+            return category;
+        }
+
+
+        public CategoryVewModelDTO GetCategoryById(int categoryId)
+        {
+            var categoryDb = _context.Categories.FirstOrDefault(x => x.CategoryId == categoryId);
+            return _mapper.Map<CategoryVewModelDTO>(categoryDb);
+        }
+
+
+        public int AddCategory(CategoryCreateDTO model)
+        {
+            var category = _mapper.Map<Category>(model);
+            _context.Categories.Add(category);
             _context.SaveChanges();
-            return categoryDb;
+            return category.CategoryId;
+        }
+
+
+        public int UpdateCategory(CategoryUpdateDTO model)
+        {
+            var isExisted = _context.Categories.Any(x => x.CategoryId == model.CategoryId);
+            if (model.ParentId!=null)
+            {
+                var isParentAExisted = _context.Categories.Any(x => x.ParentId == model.ParentId);
+                if (!isParentAExisted)
+                    return -1;
+            }
+            if (!isExisted)
+                return -1;
+            var category = _mapper.Map<Category>(model);
+            _context.Categories.Update(category);
+            _context.SaveChanges();
+            return category.CategoryId;
         }
 
         public int DeleteCategory(int id)
         {
-            var category = GetCategoryById(id);
+            var category = _context.Categories.FirstOrDefault(x => x.CategoryId == id);
+            if (category == null)
+                return -1;
             _context.Categories.Remove(category);
             _context.SaveChanges();
             return id;
